Integrate saving into map editor
This commit allows the map editor to save the map to the data file from
which it was loaded.

The editor isn't *fully* integrated, but it can now be activated for the
currently loaded map in the `game` scene by pressing E.

TODO: Implement a main menu where you can either load the map into the
game or, in development, edit said map.