Improve screen edge collision
The edge of the aircraft sprite now bounces off the edge rather than the
center (but if the aircraft is turned off-screen while still partially
off-screen, the center bounces as before). Additionally, I've made it
impossible to turn an aircraft off-screen permanently by turning it away
before it has entered the screen.